[jlm] change all float to double
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@51009 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ retex/renderer-android/src/main/java/com/himamis/retex/renderer/android/font/FontLoaderA.java
@@ -19,7 +19,7 @@ public class FontLoaderA implements FontLoader {
 	public Font loadFont(Object fontType, String name) throws ResourceParseException {
 		// TODO fontType should be a class object instead of inputstream
 		Typeface typeface = Typeface.createFromAsset(mAssetManager, BaseObjectHelper.getPath(fontType, name));
-		return new FontA(name, typeface, Math.round(PIXELS_PER_POINT));
+		return new FontA(name, typeface, (int)Math.round(PIXELS_PER_POINT));
 	}
 
 }
